fix(chart): 处理评审反馈
- pickTimeMarkerInterval 拆到独立文件,组件文件只导出组件以恢复
  Fast Refresh
- 切换谱面/难度时重置语句列表的懒解析与展开状态,收起时不为新
  谱面解析买单
- 强制 reflow 改用 getBoundingClientRect(),消除 no-unused-expressions
- renderTouchFireworks 注明入参须已按 hasFirework 过滤
